requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
ecdsa==0.19.0
fastapi==0.115.6
h11==0.14.0
httptools==0.6.4
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.3
websockets==14.1
